process command line arguments as a separate module in framework
